version complet

+++ app/Http/Controllers/MatiereController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Matiere;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class MatiereController extends Controller
 {
@@ -12,15 +13,40 @@ class MatiereController extends Controller
      */
     public function index()
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     $matieres = Matiere::all();
     return view('matieres.index', compact('matieres'));
 }
+public function mesMatieres()
+{
+    if (!Gate::allows('is-etudiant')) {
+        abort(403, 'Unauthorized action.');
+    }
+    $user = auth()->user();
+   
+
+    // Vérifiez si l'utilisateur est un étudiant
+    // if ($user->role->nom != 'etudiant') {
+    //     return redirect()->back()->withErrors(['error' => 'Cette section est réservée aux étudiants.']);
+    // }
+    // dd($user->role->nom);
+    // Récupérez les matières associées à cet étudiant
+    $matieres = $user->matieres;
+
+    return view('matieres.etudiant', compact('matieres'));
+}
+
 
     /**
      * Show the form for creating a new resource.
      */
     public function create()
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     return view('matieres.create');
 }
 
@@ -29,6 +55,9 @@ class MatiereController extends Controller
      */
     public function store(Request $request)
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     $request->validate([
         'name' => 'required|string|max:255',
         'description' => 'nullable|string',
@@ -55,6 +84,9 @@ class MatiereController extends Controller
      */
     public function edit($id)
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     $matiere = Matiere::findOrFail($id);
     return view('matieres.edit', compact('matiere'));
 }
@@ -64,12 +96,15 @@ class MatiereController extends Controller
      */
     public function update(Request $request, $id)
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     $request->validate([
-        'nom' => 'required|string|max:255',
+        'name' => 'required|string|max:255',
     ]);
 
     $matiere = Matiere::findOrFail($id);
-    $matiere->update(['nom' => $request->nom]);
+    $matiere->update(['name' => $request->name,'description' => $request->description]);
 
     return redirect()->route('matieres.index')->with('success', 'Matière mise à jour.');
 }
@@ -79,6 +114,9 @@ class MatiereController extends Controller
      */
     public function destroy($id)
 {
+    if (!Gate::allows('is-admin')) {
+        abort(403, 'Unauthorized action.');
+    }
     $matiere = Matiere::findOrFail($id);
     $matiere->delete();
